Update : take a part whole functions
