+++ includes/mini_logic.h
@@ -6,7 +6,7 @@
 /*   By: haryu <haryu@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/08 14:39:45 by haryu             #+#    #+#             */
-/*   Updated: 2022/06/17 22:33:15 by haryu            ###   ########.fr       */
+/*   Updated: 2022/06/18 13:34:59 by haryu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,8 +56,10 @@ void	*malloc_wrap(size_t size);
 char	*current_prompt(void);
 char	*ft_getcwd(void);
 int		ft_unlink(char *installed);
-int		welchs(void);
+int		welchs(char *installed);
 char	*omitted_dir(char *dir);
+char	*make_current_dir(char ***omitted);
+char	*omit_longstr(char	*str);
 char	*make_current_dir(char ***omitted);
 char	*omit_longstr(char	*str);
 int		pre_error_check(char *line);
@@ -80,10 +82,31 @@ void	free_heredoc_part(t_flist **heredoc);
 void	print_chunks(char **chunks, int height);
 void	print_heredoc_lst(t_flist **target, int height);
 void	fork_heredoc(t_flist **heredoc, int height, char *installed);
+void	clean_global(void);
+void	handler_heredoc(int signum);
+char	*make_filename(int number_cmd, int index_cmd, char *dir);
+int		make_temp(char *directory, int *cmdnum);
+void	reset_integer_vector(int **vector, int length);
+void	check_delimiter(int **cmd, t_flist *delimiter, \
+t_flist **heredoc, int maxlen);
+void	init_delimiter(int **cmd, t_flist **heredoc);
+t_flist	*find_delimiter(int **cmd, t_flist **heredoc);
+int		readline_heredoc(int fd, int **cmd, t_flist **heredoc, int height);
+void	child_heredoc(t_flist **heredoc, int height, char *installed);
+void	putin_to_temp(char *line, int *len, int fd);
 
 /* sentence part */
 
 int		sentence_part(char *line);
+int		fork_cmds(int height, char **chunks, int **pipes);
+int		ft_wait(pid_t *childs, int numbers);
+int		kill_them_all(pid_t *childs, int numbers);
+pid_t	*init_pids(int numbers);
+int		final_print(int **pipe, int height);
+void	close_pipe(int **pipe, int height);
+int		**init_pipe(int cmdnum);
+void	prepare_pipe(int ***pipes, int height);
+void	childs_test(int height, char **chunks, int **pipes);
 
 /* command part */
