+++ src/here_make_filename.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   pre_heredoc.c                                      :+:      :+:    :+:   */
+/*   here_make_filename.c                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: haryu <haryu@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/06/17 22:03:15 by haryu             #+#    #+#             */
-/*   Updated: 2022/06/18 13:14:08 by haryu            ###   ########.fr       */
+/*   Created: 2022/06/18 12:09:01 by haryu             #+#    #+#             */
+/*   Updated: 2022/06/18 12:09:17 by haryu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,17 +15,17 @@
 
 extern t_global	g_global;
 
-t_flist	**pre_heredoc(char **chunks, int height, int *heredocnum)
+char	*make_filename(int number_cmd, int index_cmd, char *dir)
 {
-	t_flist	**ret;
-	int		i;
+	char	*temp1;
+	char	*temp2;
 
-	i = 0;
-	init_flist(&ret, height);
-	while (i < height)
-	{
-		make_heredoc(chunks[i], &ret[i], heredocnum);
-		i++;
-	}
-	return (ret);
+	temp1 = ft_strjoin(dir, TEMPFILE);
+	temp2 = ft_strjoin(temp1, ft_itoa(number_cmd));
+	free(temp1);
+	temp1 = ft_strjoin(temp2, "_");
+	free(temp2);
+	temp2 = ft_strjoin(temp1, ft_itoa(index_cmd));
+	free(temp1);
+	return (temp2);
 }
