+++ src/here_putin_to_temp.c
@@ -1,24 +1,27 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_getcwd.c                                        :+:      :+:    :+:   */
+/*   here_putin_to_temp.c                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: haryu <haryu@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/06/17 21:44:12 by haryu             #+#    #+#             */
-/*   Updated: 2022/06/17 21:59:36 by haryu            ###   ########.fr       */
+/*   Created: 2022/06/18 13:08:38 by haryu             #+#    #+#             */
+/*   Updated: 2022/06/18 13:08:42 by haryu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
+#include "../includes/mini_logic.h"
 
 extern t_global	g_global;
 
-char	*ft_getcwd(void)
+void	putin_to_temp(char *line, int *len, int fd)
 {
-	char	*buf;
+	char	*putin;
 
-	buf = (char *)malloc_wrap(sizeof(char) * BUFFER);
-	getcwd(buf, BUFFER);
-	return (buf);
+	putin = NULL;
+	putin = ft_strjoin(line, "\n");
+	*len += ft_strlen(putin);
+	write(fd, putin, ft_strlen(putin));
+	free(putin);
 }
