+++ src/here_reset_integer_vector.c
@@ -1,25 +1,26 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   chunk_free.c                                       :+:      :+:    :+:   */
+/*   here_reset_integer_vector.c                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: haryu <haryu@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/06/03 03:22:28 by haryu             #+#    #+#             */
-/*   Updated: 2022/06/10 21:02:42 by haryu            ###   ########.fr       */
+/*   Created: 2022/06/18 12:56:08 by haryu             #+#    #+#             */
+/*   Updated: 2022/06/18 13:14:23 by haryu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
+#include "../includes/mini_logic.h"
 
-void chunk_free(char **chunk, size_t height)
+extern t_global	g_global;
+
+void	reset_integer_vector(int **vector, int length)
 {
-	size_t	i;
+	int	i;
 
-	i = 0;
-	while (i < height)
-		free(chunk[i++]);
-	free(chunk[i]);
-	free(chunk);
+	i = -1;
+	while (++i < length)
+		(*vector)[i] = 0;
 	return ;
 }
