+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: haryu <haryu@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/02 23:36:41 by haryu             #+#    #+#             */
-/*   Updated: 2022/06/17 22:34:09 by haryu            ###   ########.fr       */
+/*   Updated: 2022/06/18 13:34:51 by haryu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,7 +44,7 @@ void	init_process(char **installed)
 {
 	(*installed) = ft_getcwd();
 	g_global.home = getenv("HOME");
-	welchs();
+	welchs(*installed);
 	return ;
 }
