+++ src/pre_chunk_free.c
@@ -1,31 +1,25 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   pre_heredoc.c                                      :+:      :+:    :+:   */
+/*   pre_chunk_free.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: haryu <haryu@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/06/17 22:03:15 by haryu             #+#    #+#             */
-/*   Updated: 2022/06/18 13:14:08 by haryu            ###   ########.fr       */
+/*   Created: 2022/06/03 03:22:28 by haryu             #+#    #+#             */
+/*   Updated: 2022/06/18 12:00:10 by haryu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
-#include "../includes/mini_logic.h"
 
-extern t_global	g_global;
-
-t_flist	**pre_heredoc(char **chunks, int height, int *heredocnum)
+void	chunk_free(char **chunk, size_t height)
 {
-	t_flist	**ret;
-	int		i;
+	size_t	i;
 
 	i = 0;
-	init_flist(&ret, height);
 	while (i < height)
-	{
-		make_heredoc(chunks[i], &ret[i], heredocnum);
-		i++;
-	}
-	return (ret);
+		free(chunk[i++]);
+	free(chunk[i]);
+	free(chunk);
+	return ;
 }
