+++ src/sentence_final_print.c
@@ -1,31 +1,33 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   pre_heredoc.c                                      :+:      :+:    :+:   */
+/*   sentence_final_print.c                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: haryu <haryu@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/06/17 22:03:15 by haryu             #+#    #+#             */
-/*   Updated: 2022/06/18 13:14:08 by haryu            ###   ########.fr       */
+/*   Created: 2022/06/18 11:38:51 by haryu             #+#    #+#             */
+/*   Updated: 2022/06/18 13:13:57 by haryu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
-#include "../includes/mini_logic.h"
 
 extern t_global	g_global;
 
-t_flist	**pre_heredoc(char **chunks, int height, int *heredocnum)
+int	final_print(int **pipe, int height)
 {
-	t_flist	**ret;
-	int		i;
+	int		cnt;
+	char	buff[3];
 
-	i = 0;
-	init_flist(&ret, height);
-	while (i < height)
+	cnt = read(pipe[height - 1][0], buff, 1);
+	if (cnt != 0)
+		write(1, buff, 1);
+	while (cnt)
 	{
-		make_heredoc(chunks[i], &ret[i], heredocnum);
-		i++;
+		cnt = read(pipe[height - 1][0], buff, 1);
+		write(1, buff, 1);
+		if (cnt < 0)
+			break ;
 	}
-	return (ret);
+	return (FALSE);
 }
