+++ src/sentence_init_pipe.c
@@ -1,31 +1,49 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   pre_heredoc.c                                      :+:      :+:    :+:   */
+/*   sentence_init_pipe.c                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: haryu <haryu@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/06/17 22:03:15 by haryu             #+#    #+#             */
-/*   Updated: 2022/06/18 13:14:08 by haryu            ###   ########.fr       */
+/*   Created: 2022/06/18 11:35:59 by haryu             #+#    #+#             */
+/*   Updated: 2022/06/18 11:57:07 by haryu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
-#include "../includes/mini_logic.h"
 
 extern t_global	g_global;
 
-t_flist	**pre_heredoc(char **chunks, int height, int *heredocnum)
+void	prepare_pipe(int ***pipes, int height)
 {
-	t_flist	**ret;
-	int		i;
+	int	i;
 
 	i = 0;
-	init_flist(&ret, height);
 	while (i < height)
 	{
-		make_heredoc(chunks[i], &ret[i], heredocnum);
+		if (pipe((*pipes)[i]) == -1)
+		{
+			printf("Pipe error : %s\n", strerror(errno));
+			g_global.last_exitcode = 1;
+		}
 		i++;
 	}
-	return (ret);
+	if (g_global.last_exitcode == 1)
+		exit (1);
+	return ;
+}
+
+int	**init_pipe(int cmdnum)
+{
+	int		**pipes;
+	int		i;
+
+	pipes = malloc_wrap(sizeof(int *) * cmdnum);
+	i = -1;
+	while (++i < cmdnum)
+		pipes[i] = malloc_wrap(sizeof(int) * 2);
+	prepare_pipe(&pipes, cmdnum);
+	if (g_global.last_exitcode == 1)
+		exit(1);
+	return (pipes);
 }
