+++ src/sentence_kill_them_all.c
@@ -1,31 +1,25 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   pre_heredoc.c                                      :+:      :+:    :+:   */
+/*   sentence_kill_them_all.c                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: haryu <haryu@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/06/17 22:03:15 by haryu             #+#    #+#             */
-/*   Updated: 2022/06/18 13:14:08 by haryu            ###   ########.fr       */
+/*   Created: 2022/06/18 11:45:50 by haryu             #+#    #+#             */
+/*   Updated: 2022/06/18 12:03:06 by haryu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
-#include "../includes/mini_logic.h"
 
 extern t_global	g_global;
 
-t_flist	**pre_heredoc(char **chunks, int height, int *heredocnum)
+int	kill_them_all(pid_t *childs, int numbers)
 {
-	t_flist	**ret;
-	int		i;
+	int	index;
 
-	i = 0;
-	init_flist(&ret, height);
-	while (i < height)
-	{
-		make_heredoc(chunks[i], &ret[i], heredocnum);
-		i++;
-	}
-	return (ret);
+	index = -1;
+	while (++index < numbers)
+		kill(childs[index], SIGKILL);
+	return (FALSE);
 }
