+++ src/sentence_sentence_part.c
@@ -1,31 +1,31 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   pre_heredoc.c                                      :+:      :+:    :+:   */
+/*   sentence_sentence_part.c                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: haryu <haryu@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/06/17 22:03:15 by haryu             #+#    #+#             */
-/*   Updated: 2022/06/18 13:14:08 by haryu            ###   ########.fr       */
+/*   Created: 2022/06/15 21:57:25 by haryu             #+#    #+#             */
+/*   Updated: 2022/06/18 12:00:33 by haryu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
-#include "../includes/mini_logic.h"
 
 extern t_global	g_global;
 
-t_flist	**pre_heredoc(char **chunks, int height, int *heredocnum)
+int	sentence_part(char *line)
 {
-	t_flist	**ret;
-	int		i;
+	char	**chunks;
+	int		chunk_height;
+	int		**pipe;
 
-	i = 0;
-	init_flist(&ret, height);
-	while (i < height)
-	{
-		make_heredoc(chunks[i], &ret[i], heredocnum);
-		i++;
-	}
-	return (ret);
+	chunks = vertical_split(line);
+	chunk_height = check_height(line);
+	pipe = init_pipe(chunk_height);
+	if (!fork_cmds(chunk_height, chunks, pipe))
+		final_print(pipe, chunk_height);
+	close(pipe[chunk_height - 1][0]);
+	chunk_free(chunks, (size_t)chunk_height);
+	return (FALSE);
 }
